+++ packages/core-js/es/regexp/flags.js
@@ -1,5 +1,5 @@
 'use strict';
 require('../../modules/es.regexp.flags');
-var getRegExpFlags = require('../../internals/get-regexp-flags-implementation');
+var getRegExpFlags = require('../../internals/regexp-get-flags');
 
 module.exports = getRegExpFlags;
