--- packages/core-js/internals/detection-of-correctness-regexp-flags.js
@@ -1,3 +1,0 @@
-'use strict';
-
-module.exports = { correct: false };
